requirements.txt
# install with 'pip install -r requirements.txt'

Flask
Flask-Script
Flask-SQLAlchemy
gunicorn
psycopg2
twilio